refactor: deduplication
