+++ apps/daemon/pkg/toolbox/process/execute.go
@@ -5,10 +5,8 @@ package process
 
 import (
 	"errors"
-	"fmt"
 	"log/slog"
 	"net/http"
-	"os"
 	"os/exec"
 	"strings"
 	"sync/atomic"
@@ -53,13 +51,7 @@ func ExecuteCommand(logger *slog.Logger) gin.HandlerFunc {
 		if request.Cwd != nil {
 			cmd.Dir = *request.Cwd
 		}
-		if len(request.Envs) > 0 {
-			envPairs := make([]string, 0, len(request.Envs))
-			for key, value := range request.Envs {
-				envPairs = append(envPairs, fmt.Sprintf("%s=%s", key, value))
-			}
-			cmd.Env = append(os.Environ(), envPairs...)
-		}
+		common.ApplyEnvs(cmd, request.Envs)
 
 		// set maximum execution time
 		var timeoutReached atomic.Bool
